make SignConnector a module-level class
